downgrade

+++ server.sh
@@ -1,17 +1,23 @@
 #!/bin/bash
 
-. ./deploy/$1/dirs.cfg
+workon atboard
+socket=/opt/nginx/sock/atboard.sock
+uwsgi_cmd="$uwsgi_bin/uwsgi -s $socket --env PYTHONPATH=$PYTHONPATH -p 4 -M -t 20 -r -C -L -d /opt/nginx/logs/atboard.uwsgi.log -w wsgi"
 
-uwsgi_cmd="$uwsgi_bin/uwsgi -s $socket -p 4 -M -t 20 -r -C -L -d ../wsgi_watchme.log wsgi"
+nginx_conf=/opt/nginx/conf/nginx.conf
 
 case $2 in
 "start")
+if [ $1 != "workday" ]
+then
 $uwsgi_cmd
-sudo $nginx_bin/nginx -c $nginx_conf
+fi
 ;;
 "stop")
-ps aux | grep '$uwsgi_cmd' | grep -v grep | awk '{system("kill -9 " $2)}'
-sudo $nginx_bin/nginx -s stop -c $nginx_conf
+if [ $1 != "workday" ]
+then
+ps aux | grep "$uwsgi_cmd" | grep -v grep | awk '{system("kill -9 " $2)}'
+fi
 ;;
 "restart")
 bash $0 $1 stop
